requirements.txt
outscraper>=4.0.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
